requirements.txt
asgiref @ file:///work/perseverance-python-buildout/croot/asgiref_1728402006833/work
certifi==2025.10.5
charset-normalizer==3.4.4
Django==4.2.7
django-cors-headers==4.3.1
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.11
mysqlclient @ file:///croot/mysqlclient_1752659525663/work
psycopg==3.2.10
psycopg-binary==3.2.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.5
setuptools==78.1.1
sqlparse @ file:///io/buildout/croot/sqlparse_1739627228514/work
typing_extensions @ file:///croot/typing_extensions_1756280817316/work
urllib3==2.5.0
wheel==0.45.1
coverage==7.11.0
factory_boy==3.3.3
pytest @ file:///home/task_175812502822361/conda-bld/pytest_1758125086457/work
pytest-cov==7.0.0
pytest-django==4.11.1
